feat(general material): added general material section, improved layout
