+++ src/components/CategoryList/CategoryList.css
@@ -1,19 +1,30 @@
 .categories-list {
     display: flex;
     flex-direction: column;
+    width: 100%;
+    margin: 0 !important;
 }
 
 .category-link {
+    width: 100%;
+    box-sizing: border-box;
     text-align: center;
     padding: 1em 1.4em;
     border-radius: 4px;
-    margin: 0em 0em 1.2em 0em;
+    margin: 0 0 1.2em 0;
     color: white;
     text-decoration: none;
     background-color: var(--light-blue);
+    transition: background-color 0.3s ease, transform 0.3s ease;
+    box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
+}
+
+.category-link:hover {
+    background-color: var(--dark-blue);
+    box-shadow: none;
 }
 
 .category-section-title {
     text-align: center;
-    font-size: clamp(1rem, 1.2rem, 1.4rem);
+    font-size: clamp(1rem, 2.5vw, 1.4rem);
 }
