+++ src/components/CategoryList/hooks/useCategoryList.tsx
@@ -1,0 +1,18 @@
+import { useEffect, useState } from "react";
+import { getCategories } from "../../../utils";
+
+export function useCategoryList () {
+    const [courses, setCourses] = useState<{ [key: string]: string }>({});
+
+    useEffect(() => {
+        async function fetchData() {
+            const data = await getCategories();
+            setCourses(data);
+        }
+        fetchData();
+    }, []);
+
+    return {
+        courses
+    }
+}
