+++ src/components/index.ts
@@ -9,3 +9,4 @@ export {SearchPanel} from './SearchPanel/SearchPanel';
 export {UploadPanel} from './UploadPanel/UploadPanel'
 export {Layout} from './Layout/Layout'
 export {CategoryList} from './CategoryList/CategoryList'
+export {GeneralMaterial} from './GeneralMaterial/GeneralMaterial'
