+++ src/index.css
@@ -11,7 +11,7 @@
   --orange: #ED6D33;
   --blue: #548AF3;
   --light-blue: #3c7dff;
-  --dark-blue: #548AF3;
+  --dark-blue: #1d5ddf;
   --purple: #C054F3;
   --light-purple: #d18df1;
   --red: #F43D3D;
