+++ src/pages/HomePage/HomePage.css
@@ -1,0 +1,12 @@
+.action-container {
+    display: flex;
+    flex-direction: row;
+    gap: 1.2em;
+}
+
+@media screen and (max-width: 768px){
+    .action-container {
+        flex-direction: column;
+        gap: 1.4em;
+    }
+}
